backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
python-dotenv
flask-jwt-extended
Flask-Mail
psycopg2-binary
gunicorn
